requirements.txt
annotated-types==0.7.0
anthropic==0.45.1
anyio==4.6.0
asgiref==3.8.1
beautifulsoup4==4.12.3
bs4==0.0.2
cachetools==5.5.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.1
colorama==0.4.6
cryptography==44.0.2
Cython==3.0.11
display==1.0.0
distro==1.9.0
Django==5.1.1
django-allauth==65.4.1
dnspython==2.7.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.162.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-genai==1.3.0
google-generativeai==0.8.4
googleapis-common-protos==1.68.0
grpcio==1.70.0
grpcio-status==1.70.0
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.2
idna==3.10
jiter==0.5.0
jwt==1.3.1
openai==1.50.2
pillow==11.1.0
proto-plus==1.26.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pymongo==4.11
pyparsing==3.2.1
python-dotenv==1.0.1
requests==2.32.3
rsa==4.9
setuptools==75.1.0
sniffio==1.3.1
soupsieve==2.6
sqlparse==0.5.1
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.3.0
websockets==14.2
